paho-services/requirements.txt
paho-mqtt
chirpstack-api
